Update `component-scoped-csr` example
