+++ package.json
@@ -46,7 +46,7 @@
   },
   "homepage": "https://github.com/sveltekit-i18n/lib#readme",
   "peerDependencies": {
-    "svelte": "^4.0.5"
+    "svelte": ">=3.49.0"
   },
   "devDependencies": {
     "@types/jest": "^29.5.2",
